# backend/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Nurse Api",
        "description": "E-Nurse Api description",
        "contact": {
            "email": "developer@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/nurses": {
            "get": {
                "tags": [
                    "Nurse"
                ],
                "summary": "Get All Nurses",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "getNurses",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Nurse"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Nurse"
                ],
                "summary": "Create Nurse",
                "description": "Multiple values can be provided with comma separated string",
                "operationId": "createNurse",
                "requestBody": {
                    "description": "Nurse to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Nurse"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nurse created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Nurse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Nurse"
                ],
                "summary": "Update Nurse",
                "description": "Update Nurse",
                "operationId": "updateNurse",
                "requestBody": {
                    "description": "Nurse to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Nurse"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nurse updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Nurse"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "404": {
                        "description": "Nurse not found."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/nurses/{id}": {
            "get": {
                "tags": [
                    "Nurse"
                ],
                "summary": "Get Nurse",
                "description": "Nurse id should be provided",
                "operationId": "getNurse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Get Nurse",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Nurse"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Nurse"
                ],
                "summary": "Delete a Nurse",
                "description": "Delete Nurse",
                "operationId": "deleteNurse",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Nurse--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Nurse not found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Nurse": {
                "title": "Nurse",
                "description": "Nurse",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Nurse id",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Nurse Name (Max:50)",
                        "type": "string",
                        "example": "ahmad"
                    },
                    "email": {
                        "title": "email",
                        "description": "Nurse Email (Should be Unique)",
                        "type": "string",
                        "example": "ahmad@example.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Nurse password (Min:8)",
                        "type": "string",
                        "example": "12345678"
                    },
                    "gender": {
                        "title": "gender",
                        "description": "Nurse gender",
                        "type": "boolean",
                        "example": "1"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Nurse phone",
                        "type": "string",
                        "example": "0999111222"
                    },
                    "is_resigned": {
                        "title": "is_resigned",
                        "description": "Resigning status",
                        "type": "boolean",
                        "example": "0"
                    },
                    "is_admin": {
                        "title": "is_admin",
                        "description": "Nurse Role (admin|user)",
                        "type": "boolean",
                        "example": "0"
                    },
                    "created_at": {
                        "title": "Create Date",
                        "description": "Create Date)",
                        "type": "string",
                        "example": "2023-04-19 12:37:32 PM"
                    },
                    "updated_at": {
                        "title": "Update Date",
                        "description": "Update Date",
                        "type": "string",
                        "example": "2023-04-19 10:37:32 AM"
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
...
            "Nurse--id": {
                "name": "id",
                "in": "path",
                "description": "Nurse Id",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64",
                    "example": "1"
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}